eccci: Fix build error with newer clang
../drivers/misc/mediatek/eccci/hif/ccci_hif_dpmaif.c:465:4: error: cast to smaller integer type 'u32' (aka 'unsigned int') from 'u64 *' (aka 'unsigned long long *') [-Werror,-Wpointer-to-int-cast]
                        (u32)data_64ptr, (i * 8),
                        ^~~~~~~~~~~~~~~
../drivers/misc/mediatek/eccci/hif/ccci_hif_dpmaif.c:121:51: note: expanded from macro 'DPMA_DRB_DATA_INFO'
        ccci_dump_write(0, CCCI_DUMP_DPMA_DRB, 0, fmt, ##args)
                                                         ^~~~
../drivers/misc/mediatek/eccci/hif/ccci_hif_dpmaif.c:478:37: error: cast to smaller integer type 'u32' (aka 'unsigned int') from 'u8 *' (aka 'unsigned char *') [-Werror,-Wpointer-to-int-cast]
                DPMA_DRB_DATA_INFO("%08X(%04d):", (u32)data_8ptr, count * 8);
                                                  ^~~~~~~~~~~~~~
../drivers/misc/mediatek/eccci/hif/ccci_hif_dpmaif.c:121:51: note: expanded from macro 'DPMA_DRB_DATA_INFO'
        ccci_dump_write(0, CCCI_DUMP_DPMA_DRB, 0, fmt, ##args)
                                                         ^~~~
2 errors generated.

+++ drivers/misc/mediatek/eccci/hif/ccci_hif_dpmaif.c
@@ -449,7 +449,7 @@ static void dump_drb_queue_data(unsigned int qno)
 	i = 0;
 	while (i < count) {
 		DPMA_DRB_DATA_INFO("%08X(%04d): %016llX %016llX %016llX %016llX %016llX %016llX %016llX %016llX\n",
-			(u32)data_64ptr, (i * 8),
+			(long)data_64ptr, (i * 8),
 			*data_64ptr, *(data_64ptr + 1),
 			*(data_64ptr + 2), *(data_64ptr + 3),
 			*(data_64ptr + 4), *(data_64ptr + 5),
@@ -462,7 +462,7 @@ static void dump_drb_queue_data(unsigned int qno)
 	if (mod64 > 0) {
 		data_8ptr = (u8 *)data_64ptr;
 
-		DPMA_DRB_DATA_INFO("%08X(%04d):", (u32)data_8ptr, count * 8);
+		DPMA_DRB_DATA_INFO("%08X(%04d):", (long)data_8ptr, count * 8);
 
 		for (i = 0; i < mod64; i++) {
 			if ((i % 8) == 0)
